# config/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "identity-verificate-service",
        "version": "1.0",
        "description": "本人性確認サービス"
    },
    "servers": [
        {
            "url": "/identity-verificate"
        }
    ],
    "paths": {
        "/code": {
            "post": {
                "tags": [
                    "本人性確認コード発行"
                ],
                "description": "本人性確認コードを発行し、WF関連の発行処理の場合はWFメンバーへの通知も行います。",
                "requestBody": {
                    "description": "本人性確認コード発行用のリクエスト",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "発行情報",
                                "properties": {
                                    "actor": {
                                        "type": "object",
                                        "properties": {
                                            "_value": {
                                                "type": "integer",
                                                "description": "カタログコード"
                                            },
                                            "_ver": {
                                                "type": "integer",
                                                "description": "カタログバージョン"
                                            }
                                        },
                                        "description": "アクターカタログコードオブジェクト"
                                    },
                                    "wf": {
                                        "type": "object",
                                        "properties": {
                                            "_value": {
                                                "type": "integer",
                                                "description": "カタログコード"
                                            },
                                            "_ver": {
                                                "type": "integer",
                                                "description": "カタログバージョン"
                                            }
                                        },
                                        "description": "WF定義カタログコードオブジェクト"
                                    },
                                    "app": {
                                        "type": "object",
                                        "properties": {
                                            "_value": {
                                                "type": "integer",
                                                "description": "カタログコード"
                                            },
                                            "_ver": {
                                                "type": "integer",
                                                "description": "カタログバージョン"
                                            }
                                        },
                                        "description": "APP定義カタログコードオブジェクト"
                                    }
                                },
                                "example": {
                                    "actor": {
                                        "_value": 1000004,
                                        "_ver": 1
                                    },
                                    "app": null,
                                    "wf": {
                                        "_value": 1000007,
                                        "_ver": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "コード発行に成功した際のレスポンス",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "発行情報",
                                    "properties": {
                                        "identifyCode": {
                                            "type": "string",
                                            "description": "本人性確認コード"
                                        },
                                        "actor": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "カタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "カタログバージョン"
                                                }
                                            },
                                            "description": "アクターカタログコードオブジェクト"
                                        },
                                        "wf": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "カタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "カタログバージョン"
                                                }
                                            },
                                            "description": "WF定義カタログコードオブジェクト"
                                        },
                                        "app": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "カタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "カタログバージョン"
                                                }
                                            },
                                            "description": "APP定義カタログコードオブジェクト"
                                        },
                                        "expireAt": {
                                            "type": "string",
                                            "description": "有効期限(yyyy-MM-ddTHH:MM:SS.SSS+0900"
                                        }
                                    }
                                },
                                "example": {
                                    "identifyCode": "ukO8z+Xf8vv7yxXQj2Hpo",
                                    "actor": {
                                        "_value": 1000004,
                                        "_ver": 1
                                    },
                                    "app": null,
                                    "wf": {
                                        "_value": 1000007,
                                        "_ver": 1
                                    },
                                    "expireAt": "2020-12-01T00:00:00.000+0900"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/code/verified": {
            "post": {
                "tags": [
                    "確認済本人性確認コード生成"
                ],
                "description": "確認済の本人性確認コードを発行します。",
                "requestBody": {
                    "description": "確認済本人性確認コード生成用のリクエスト",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "確認コード生成情報",
                                "properties": {
                                    "pxrId": {
                                        "type": "string",
                                        "description": "連携PXR-ID"
                                    },
                                    "userId": {
                                        "type": "string",
                                        "description": "連携利用者ID"
                                    },
                                    "actor": {
                                        "type": "object",
                                        "properties": {
                                            "_value": {
                                                "type": "integer",
                                                "description": "連携アクターカタログコード"
                                            },
                                            "_ver": {
                                                "type": "integer",
                                                "description": "連携アクターカタログバージョン"
                                            }
                                        },
                                        "description": "アクターカタログコードオブジェクト"
                                    },
                                    "wf": {
                                        "type": "object",
                                        "properties": {
                                            "_value": {
                                                "type": "integer",
                                                "description": "連携wfカタログコード"
                                            },
                                            "_ver": {
                                                "type": "integer",
                                                "description": "連携wfカタログバージョン"
                                            }
                                        },
                                        "description": "WF定義カタログコードオブジェクト"
                                    },
                                    "app": {
                                        "type": "object",
                                        "properties": {
                                            "_value": {
                                                "type": "integer",
                                                "description": "連携appカタログコード"
                                            },
                                            "_ver": {
                                                "type": "integer",
                                                "description": "連携appカタログバージョン"
                                            }
                                        },
                                        "description": "APP定義カタログコードオブジェクト"
                                    }
                                },
                                "example": {
                                    "pxrId": "test.org",
                                    "userId": "test.bank",
                                    "actor": {
                                        "_value": 1000004,
                                        "_ver": 1
                                    },
                                    "app": null,
                                    "wf": {
                                        "_value": 1000007,
                                        "_ver": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "確認済本人性確認コード生成に成功した際のレスポンス",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "確認コード生成情報",
                                    "properties": {
                                        "identifyCode": {
                                            "type": "string",
                                            "description": "本人性確認コード"
                                        },
                                        "pxrId": {
                                            "type": "string",
                                            "description": "連携PXR-ID"
                                        },
                                        "userId": {
                                            "type": "string",
                                            "description": "連携利用者ID"
                                        },
                                        "actor": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "連携アクターカタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "連携アクターカタログバージョン"
                                                }
                                            },
                                            "description": "アクターカタログコードオブジェクト"
                                        },
                                        "wf": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "連携wfカタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "連携wfカタログバージョン"
                                                }
                                            },
                                            "description": "WF定義カタログコードオブジェクト"
                                        },
                                        "app": {
                                            "type": "object",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "連携appカタログコード"
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "連携appカタログバージョン"
                                                }
                                            },
                                            "description": "APP定義カタログコードオブジェクト"
                                        },
                                        "expireAt": {
                                            "type": "string",
                                            "description": "有効期限(yyyy-MM-ddTHH:MM:SS.SSS+0900"
                                        }
                                    }
                                },
                                "example": {
                                    "identifyCode": "ukO8z+Xf8vv7yxXQj2Hpo",
                                    "pxrId": "test.org",
                                    "userId": "test.bank",
                                    "actor": {
                                        "_value": 1000004,
                                        "_ver": 1
                                    },
                                    "app": null,
                                    "wf": {
                                        "_value": 1000007,
                                        "_ver": 1
                                    },
                                    "expireAt": "2020-12-01T00:00:00.000+0900"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}